+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog of elbencho
 
+## v1.6.0 (under construction)
+
+### New Features & Enhancements
+* New option to truncate files to full size (--trunctosize) in creation/write phase.
+* New option to preallocate file disk space (--preallocfile) in creation/write phase.
+
 ## v1.5.3 (Nov 15, 2020)
 
 ### Fixes